RequestService I added getPendingRegistration, get PendingInternshipVet, getPendingWithdrawal.
StaffController I implemented launch method, that initialises a new StaffUI and passes SystemController and StaffController instances into it, idk how else to get access to them.

StaffController I also implemented some new methods that give me Pending requests and return the staff object.

StaffUI I implemented some basic logic for each function. Not very sure about the changePassword and UpdateFilter.

For SystemController in the handleLogin, for the case of staff I passed this into launch so as to be able to pass it into the StaffUI instance.